+++ src/adapter/bun/index.ts
@@ -64,14 +64,11 @@ const getPossibleParams = (path: string) => {
 	return routes
 }
 
-export const isHTMLBundle = (handle: any) => {
-	return (
-		typeof handle === 'object' &&
-		handle !== null &&
-		(handle.toString() === '[object HTMLBundle]' ||
-			typeof handle.index === 'string')
-	)
-}
+export const isHTMLBundle = (handle: any) =>
+	typeof handle === 'object' &&
+	handle !== null &&
+	(handle.toString() === '[object HTMLBundle]' ||
+		typeof handle.index === 'string')
 
 const supportedMethods = {
 	GET: true,